requirements.txt
pygithub==1.26.0
stemming==1.0.1
nltk